refactored, added index files for components
