+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
-import { Navbar } from './components/Navbar/Navbar';
-import { NotFoundPage } from './components/NotFoundPage/NotFoundPage';
-import { HomePage } from './components/HomePage/HomePage';
-import { PeoplePage } from './components/PeoplePage/PeoplePage';
+import { Navbar } from './components/Navbar';
+import { NotFoundPage } from './components/NotFoundPage';
+import { HomePage } from './components/HomePage';
+import { PeoplePage } from './components/PeoplePage';
 
 export const App = () => (
   <div data-cy="app">
@@ -12,9 +12,9 @@ export const App = () => (
       <div className="container">
         <Routes>
           <Route path="/" element={<HomePage />} />
-          <Route path="/home" element={<Navigate to="/" replace />} />
-          <Route path="/people" element={<PeoplePage />}>
-            <Route path="/people/:slug" element={<PeoplePage />} />
+          <Route path="home" element={<Navigate to="/" replace />} />
+          <Route path="people" element={<PeoplePage />}>
+            <Route path=":slug" element={<PeoplePage />} />
           </Route>
           <Route path="*" element={<NotFoundPage />} />
         </Routes>
